Extract out of order records tests to a separate class
