+++ java/core/src/test/java/sleeper/core/record/process/status/ProcessRunsTest.java
@@ -24,13 +24,14 @@ import java.time.Instant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.groups.Tuple.tuple;
 import static sleeper.core.record.process.status.TestProcessRuns.runsFromUpdates;
+import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.DEFAULT_TASK_ID;
+import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.TASK_ID_1;
+import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.TASK_ID_2;
 import static sleeper.core.record.process.status.TestProcessStatusUpdateRecords.onTask;
 import static sleeper.core.record.process.status.TestRunStatusUpdates.finishedStatus;
 import static sleeper.core.record.process.status.TestRunStatusUpdates.startedStatus;
 
 public class ProcessRunsTest {
-    private static final String TASK_ID_1 = "task-id-1";
-    private static final String TASK_ID_2 = "task-id-2";
 
     @Test
     public void shouldReportNoFinishedStatusWhenJobNotFinished() {
@@ -44,7 +45,7 @@ public class ProcessRunsTest {
         assertThat(runs.getRunList())
                 .extracting(ProcessRun::getTaskId, ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
                 .containsExactly(
-                        tuple(TASK_ID_1, started, null));
+                        tuple(DEFAULT_TASK_ID, started, null));
         assertThat(runs.isFinished()).isFalse();
     }
 
@@ -61,7 +62,7 @@ public class ProcessRunsTest {
         assertThat(runs.getRunList())
                 .extracting(ProcessRun::getTaskId, ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
                 .containsExactly(
-                        tuple(TASK_ID_1, started, finished));
+                        tuple(DEFAULT_TASK_ID, started, finished));
         assertThat(runs.isFinished()).isTrue();
     }
 
@@ -78,8 +79,8 @@ public class ProcessRunsTest {
         assertThat(runs.getRunList())
                 .extracting(ProcessRun::getTaskId, ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
                 .containsExactly(
-                        tuple(TASK_ID_1, started2, null),
-                        tuple(TASK_ID_1, started1, null));
+                        tuple(DEFAULT_TASK_ID, started2, null),
+                        tuple(DEFAULT_TASK_ID, started1, null));
         assertThat(runs.isFinished()).isFalse();
     }
 
@@ -98,8 +99,8 @@ public class ProcessRunsTest {
         assertThat(runs.getRunList())
                 .extracting(ProcessRun::getTaskId, ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
                 .containsExactly(
-                        tuple(TASK_ID_1, started2, finished2),
-                        tuple(TASK_ID_1, started1, finished1));
+                        tuple(DEFAULT_TASK_ID, started2, finished2),
+                        tuple(DEFAULT_TASK_ID, started1, finished1));
         assertThat(runs.isFinished()).isTrue();
     }
 
@@ -128,7 +129,7 @@ public class ProcessRunsTest {
     }
 
     @Test
-    public void shouldReportTwoTasksWithOneFinishedRunEachForSameJob() {
+    public void shouldReportTwoTasksWithOneFinishedRunEach() {
         // Given
         ProcessStartedStatus started1 = startedStatus(Instant.parse("2022-09-23T09:23:30.001Z"));
         ProcessFinishedStatus finished1 = finishedStatus(started1, Duration.ofSeconds(30), 450L, 300L);
@@ -139,86 +140,6 @@ public class ProcessRunsTest {
         ProcessRuns runs = runsFromUpdates(
                 onTask(TASK_ID_1, started1, finished1),
                 onTask(TASK_ID_2, started2, finished2));
-
-        // Then
-        assertThat(runs.getRunList())
-                .extracting(ProcessRun::getTaskId, ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
-                .containsExactly(
-                        tuple(TASK_ID_2, started2, finished2),
-                        tuple(TASK_ID_1, started1, finished1));
-        assertThat(runs.isFinished()).isTrue();
-    }
-
-    @Test
-    public void shouldReportRunWhenJobFinishedReturnedFromDatabaseOutOfOrder() {
-        // Given
-        ProcessStartedStatus started = startedStatus(Instant.parse("2022-09-24T09:23:30.001Z"));
-        ProcessFinishedStatus finished = finishedStatus(started, Duration.ofSeconds(30), 450L, 300L);
-
-        // When
-        ProcessRuns runs = runsFromUpdates(finished, started);
-
-        // Then
-        assertThat(runs.getRunList())
-                .extracting(ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
-                .containsExactly(
-                        tuple(started, finished));
-        assertThat(runs.isFinished()).isTrue();
-    }
-
-    @Test
-    public void shouldReportRunsWhenJobStartedReturnedFromDatabaseOutOfOrder() {
-        // Given
-        ProcessStartedStatus started1 = startedStatus(Instant.parse("2022-09-24T09:23:30.001Z"));
-        ProcessStartedStatus started2 = startedStatus(Instant.parse("2022-09-25T09:23:30.001Z"));
-        ProcessStartedStatus started3 = startedStatus(Instant.parse("2022-09-26T09:23:30.001Z"));
-
-        // When
-        ProcessRuns runs = runsFromUpdates(started3, started1, started2);
-
-        // Then
-        assertThat(runs.getRunList())
-                .extracting(ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
-                .containsExactly(
-                        tuple(started3, null),
-                        tuple(started2, null),
-                        tuple(started1, null));
-        assertThat(runs.isFinished()).isFalse();
-    }
-
-    @Test
-    public void shouldReportRunsWhenLastRunFinishedButReturnedFromDatabaseOutOfOrder() {
-        // Given
-        ProcessStartedStatus started1 = startedStatus(Instant.parse("2022-09-24T09:23:30.001Z"));
-        ProcessStartedStatus started2 = startedStatus(Instant.parse("2022-09-25T09:23:30.001Z"));
-        ProcessStartedStatus started3 = startedStatus(Instant.parse("2022-09-26T09:23:30.001Z"));
-        ProcessFinishedStatus finished = finishedStatus(started3, Duration.ofSeconds(30), 450L, 300L);
-
-        // When
-        ProcessRuns runs = runsFromUpdates(started3, finished, started1, started2);
-
-        // Then
-        assertThat(runs.getRunList())
-                .extracting(ProcessRun::getStartedStatus, ProcessRun::getFinishedStatus)
-                .containsExactly(
-                        tuple(started3, finished),
-                        tuple(started2, null),
-                        tuple(started1, null));
-        assertThat(runs.isFinished()).isFalse();
-    }
-
-    @Test
-    public void shouldReportRunsOnDifferentTasksWhenJobFinishedFromDatabaseOutOfOrder() {
-        // Given
-        ProcessStartedStatus started1 = startedStatus(Instant.parse("2022-09-22T09:23:30.001Z"));
-        ProcessFinishedStatus finished1 = finishedStatus(started1, Duration.ofSeconds(30), 450L, 300L);
-        ProcessStartedStatus started2 = startedStatus(Instant.parse("2022-09-22T09:23:31.001Z"));
-        ProcessFinishedStatus finished2 = finishedStatus(started2, Duration.ofSeconds(30), 450L, 300L);
-
-        // When
-        ProcessRuns runs = runsFromUpdates(
-                onTask(TASK_ID_1, finished1, started1),
-                onTask(TASK_ID_2, finished2, started2));
 
         // Then
         assertThat(runs.getRunList())
